+++ FFT/TPCircularBuffer.c
@@ -37,7 +37,7 @@ static inline bool _checkResult(kern_return_t result, const char *operation, con
 
 bool TPCircularBufferInit(TPCircularBuffer *buffer, int length) {
     
-    buffer->length = round_page(length);    // We need whole page sizes
+    buffer->length = (int32_t)round_page(length);    // We need whole page sizes
     buffer->count = 0;
     // Temporarily allocate twice the length, so we have the contiguous address space to
     // support a second instance of the buffer directly after
